+++ docs/source/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.3.1
+-----
+
+- Improved and fixed error codes in documentation
+- Added the :py:func`aito.schema.AitoColumnTypeSchema.infer_from_samples` function that infers the Column Type from samples.
+
 0.3.0
 -----